fix: restrict gridSize to number type
In bringing the code over from CodePen, discovered that the code here
does not recognize the gridSize variable as a type without explicitly
forcing it to be so. This ensures that this variable will always be a
number.
*Since this fix was made using `git add -p`, a console.log removal
also sneaked in where git would not allow me to commit a smaller chunk.

FSA2021-104