feat: set --size in JavaScript
-In order to maintain perfect squares in the grid, we need to explicitly
set height and width of each tile, but since grid size can vary, we need
to have JavaScript tell CSS about the grid size that has been set in the
form. Here, we set the --size variable for our CSS in JavaScript, then
refer to it in our CSS to create a width/height for our tiles based on
viewport width.

FSA2021-118