feat: add contrast background to checkmark
The white checkmark on the colored tile has been changed to a black
checkmark with a white circular background. Also, now the checkmarks/
checkmark backgrounds are sized to fit the content using the same --grid
custom property as what's used to size the tiles.
Additionally: to be more realistic about what kinds of grid sizes we
might be working with, the grid size has been restricted to between 3
and 16.
NOTE: We are having issues with getting the values of the custom
properties in Chrome DevTools--this appears to be a Chrome-specific
issue, and after some assessment, we are confident that the styles are
being computed properly despite not being defined in Chrome.

FSA2021-167