feat: add game and sandbox buttons/modes
1: Add markup for header and navigation:
-In order to add separate game and sandbox modes, we must have a way to
navigate between the two. This commit adds a nav element containing the
buttons that will switch the mode of the game.
-Adding a basic header to the markup as well so that the navigation
isn't the first thing on the page.

2: Remove form functions from index.js:
-In order to allow index.js to serve as an index/entry point, all of the
previous functions that dealt with submitting the form have been moved
out and replaced with a function call to addFormListener() which will
call all of the same form functions that were previously in index.js.
-The clearGridContainer() function has been moved out to dom-helpers as
it seemed to fit better there.

3: Add separate app modes
-The game will start in sandbox mode but we want the option to switch to
game mode as we like. When we start the app, it will now call a function
initSandboxMode() that does all of the things that we want to happen in
sandbox mode and it will add listeners to our navigation so that we can
toggle between game and sandbox mode.
-When we click on the game mode button, it will call a function called
initGameMode() that will do all of the things that we want to happen in
game mode.

4: Add visible differences between game and sandbox mode:
-Shows form if in sandbox mode, hides form if in game mode
-Remove any existing grid when navigating back to sandbox mode from game
-Remove play button and winner header when in sandbox mode
-When going back to game mode, it starts with a new grid

5: Use default grid if there is no grid:
-If there is no grid created when game mode is clicked, uses the grid
based on whatever the form's current defaults are.

FSA2021-149