refactor: change game flow
In order to implement UX suggestions, the flow of the app was reworked.

Includes:
-Updating header and styling text
-Adding paragraph text to guide the user
-Removing original nav and replacing it
-Styling buttons with default styles
-Centering the form
-Changing number input from text input to slider
-Changing the initial behavior of the app
-Adding navigation with buttons that are relevant for the game state
-Adding reusable functions to handle various elements of the game

However, in doing this, I created a dependency cycle that will need to
be handled later on. The way that the app has been architected, we have created dependency
cycles, most often when adding a callback function for an event
listener. This SHOULD be fixed at some point later in the future, but
for now, this eslint rule will be disabled.

FSA2021-148