feat: disable get colors button until colors are picked
-Disables get colors button until at least 3 of the color inputs are
unique/different from each other.
-When we go back to the grid (using the change grid button), the get
colors button was disabled even though there are 3-4 unique colors
in the color inputs. Now the button is only disabled at the beginning
of the app and when starting over from the beginning. To accomplish
this, the entry point from index.js is now restartApp instead of
startApp.
-Note: This validation is only turned on when starting the app, it does
not validate again after it's been validated once unless someone
clicks on the "New Game" button.

FSA2021-150